requirements.txt
google-cloud-vision==0.39.0
google-cloud-storage==1.19.1
ffmpeg-python
